sc - restructure folders to follow page + component structure
