--- frontend/src/.env.production.SAMPLE
@@ -1,2 +1,0 @@
-REACT_APP_AUTH0_DOMAIN=FILL-IT-IN
-REACT_APP_AUTH0_CLIENT_ID=FILL-IT-IN
