+++ frontend/src/components/Footer/AppFooter.test.js
