--- frontend/src/components/auth0/index.js
@@ -1,3 +1,0 @@
-import Auth0ProviderWithHistory from "./Auth0ProviderWithHistory";
-
-export { Auth0ProviderWithHistory };
